+++ MyWallet.Domain/Entities/Ticker.cs
@@ -3,7 +3,7 @@
     public class Ticker : Entity
     {
         public string Code { get; private set; }
-        public int CompanyId { get; private set; }
+        public long CompanyId { get; private set; }
         public decimal? Price { get; private set; }
         public DateTime? UpdateDate { get; private set; }
         public virtual Company Company { get; private set; }
